Apply CodeRabbit feedback: Reuse the findFencedBlock function
